requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
distro==1.9.0
et-xmlfile==1.1.0
Flask==2.0.1
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.4
numpy==1.25.0
openai==1.14.1
openpyxl==3.1.2
pandas==2.0.2
pandas-stubs==2.0.2.230605
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
tqdm==4.65.0
types-pytz==2023.3.0.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==3.0.1
yarl==1.9.2
